cleanup config.py and move to experiments
- Also removes dependency on config module from platalea and utils modules.
- config.device() is now hardware.device() in platalea.
- Note that evaluate_asr and extract_transcriptions in utils do not work in this commit, will be fixed in next when we move the utils and experiments folders into platalea to make them subpackages.
