+++ platalea/asr.py
@@ -13,9 +13,9 @@ from platalea.decoders import TextDecoder
 from platalea.encoders import SpeechEncoder
 import platalea.loss
 import platalea.score
-import platalea.config
+import platalea.hardware
 
-_device = platalea.config.device()
+_device = platalea.hardware.device()
 
 
 class SpeechTranscriber(nn.Module):
