+++ platalea/basic.py
@@ -11,10 +11,9 @@ from platalea.encoders import SpeechEncoder, ImageEncoder
 import platalea.loss
 import platalea.dataset as D
 import platalea.score
-import platalea.config
+import platalea.hardware
 
-
-_device = platalea.config.device()
+_device = platalea.hardware.device()
 
 
 class SpeechImage(nn.Module):
