+++ platalea/mtl.py
@@ -12,9 +12,9 @@ from platalea.speech_text import SpeechText
 from platalea.asr import SpeechTranscriber
 import platalea.loss
 import platalea.score
-import platalea.config
+import platalea.hardware
 
-_device = platalea.config.device()
+_device = platalea.hardware.device()
 
 
 class MTLNetASR(nn.Module):
