+++ platalea/preprocessing.py
@@ -7,16 +7,13 @@ import json
 import logging
 from scipy.io.wavfile import read
 import numpy
-import platalea.config
 import pathlib
+import platalea.hardware
+
+_device = platalea.hardware.device()
 
 
-_device = platalea.config.device()
-
-
-def flickr8k_features(dataset_path=platalea.config.args.data_root,
-                      audio_subdir=platalea.config.args.audio_subdir,
-                      images_subdir=platalea.config.args.image_subdir):
+def flickr8k_features(dataset_path, audio_subdir, images_subdir):
     audio_config = dict(dataset_path=pathlib.Path(dataset_path), audio_subdir=audio_subdir, type='mfcc',
                         delta=True, alpha=0.97, n_filters=40, window_size=0.025,
                         frame_shift=0.010)
