+++ platalea/text_image.py
@@ -11,9 +11,9 @@ import platalea.dataset as D
 from platalea.encoders import TextEncoder, ImageEncoder
 import platalea.loss
 import platalea.score
-import platalea.config
+import platalea.hardware
 
-_device = platalea.config.device()
+_device = platalea.hardware.device()
 
 
 class TextImage(nn.Module):
